added ARM11 MMU from skyeye

+++ src/core/src/mem_map.h
@@ -71,8 +71,8 @@ u32 Read32(const u32 addr);
 u32 Read8_ZX(const u32 addr);
 u32 Read16_ZX(const u32 addr);
 
-void Write8(const u32 addr, const u32 data);
-void Write16(const u32 addr, const u32 data);
+void Write8(const u32 addr, const u8 data);
+void Write16(const u32 addr, const u16 data);
 void Write32(const u32 addr, const u32 data);
 
 u8* GetPointer(const u32 Address);
